requirements.txt
aioice==0.10.2
aiortc==1.14.0
altair==6.0.0
attrs==25.4.0
av==16.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
dlib==20.0.0
dnspython==2.8.0
face-recognition==1.3.0
face_recognition_models @ git+https://github.com/ageitgey/face_recognition_models@e67de717267507d1e9246de95692eb8be736ab61
gitdb==4.0.12
GitPython==3.1.46
google-crc32c==1.8.0
idna==3.11
ifaddr==0.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.16.0
numpy==2.2.6
opencv-python==4.13.0.92
packaging==26.0
pandas==2.3.3
pillow==12.1.0
protobuf==6.33.5
pyarrow==23.0.0
pycparser==3.0
pydeck==0.9.1
pyee==13.0.0
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
streamlit==1.54.0
streamlit-webrtc==0.64.5
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
